Change Queue End
After this change. You will be able to play songs after queue end. Before this if the queue end it didn't play a new song against this it add song to the queue which was deleted after leaveOnEnd time and never played a new song after skipping the last song from the queue.